Openapi spec timeline (#163083)
## Summary

We are working to implement our public api's in the openapi standard
format

## Testing

Either:
1. Install an editor extension such as:
https://marketplace.visualstudio.com/items?itemName=42Crunch.vscode-openapi
and preview as is

<img width="1715" alt="image"
src="https://github.com/elastic/kibana/assets/17211684/23f7ebcc-f10c-452f-8456-1f8e8e38545e">

Or if you would rather look online: https://editor-next.swagger.io/
- The only issue is the reference to the `components.yaml`. To get
around this, copy the entire components yaml to the bottom of any of the
route_schema.yaml files and update the references by removing the first
path reference of `'../model/components.yaml` as such:
- `$ref: '../model/components.yaml#/components/schemas/TimelineStatus'`
=>
    - `$ref: '#/components/schemas/TimelineStatus'`

### Notes

- `resolve/` api. I wonder if this should remain internal as the point
of this api was to resolve 7.x savedObjectIds to their 8.x counterparts
after the upgrade. This is primarily used for maintaining the
functionality for links with so id references copied in 7.x, so that
when a user clicks it, it will resolve properly to the 8.x SO.

### Questions
- Should underscored paths actually just be made private?
i.e. `import | export | draft | favorite | prepackaged` @angorayc or
@XavierM may have insight here

### Tracking

| Done? | Method | Endpoint |
| ----- | ------ |
------------------------------------------------------- |
| yes | POST    | /api/timeline       |
| yes | PATCH    | /api/timeline       |
| yes | POST    | /api/timeline/_import       |
| yes | POST    | /api/timeline/_export       |
| yes | GET    | /api/timeline/_draft       |
| yes | GET    | /api/timeline       |
| TODO? | GET    | /api/timeline/resolve       |
| yes | GET    | /api/timelines       |
| yes | POST    | /api/timeline/_draft       |
| yes | DELETE    | /api/timeline       |
| yes | PATCH    | /api/timeline/_favorite       |
| yes | POST    | /api/timeline/_prepackaged       |
| yes | PATCH    | /api/note       |
| yes | DELETE    | /api/note       |
| yes | PATCH    | /api/pinned_event       |

(cherry picked from commit 2c7741552ac7157ee12623a460de28e5202197f1)