Use a more robust SavedObjects client in our risk scoring task
Rather than defining the SO extensions ourselves, we instead build a
fake kibana request for the particular namespace we're in, and use the
request-based helpers to build a scoped SO client from that request. In
this way, we're not coupled to the extensions implementation, and
additionally we have the encryption and spaces extensions
defined/enabled for us.

We need to disable the security extension, however, because there is not
a kibana system user that would be able to access these Saved Objects,
normally.