[ML] Removing token list from text expansion model testing (#164560)
The tokens listed could contain sensitive words which we do not want to
display to the user, in case they cause offence.
For now we can just hide this list, in case a future version of the API
contains a sanitised list of tokens.

(cherry picked from commit 323878e746116550b23f25269b9d65d86552fc03)