[8.9] [Security Solution] Fix flaky test: detection_rules/bulk_edit_rules_actions.cy.ts (#164569)
**NOTE: This is a manual backport of
https://github.com/elastic/kibana/pull/163698**

**Original PR description:**

Fixes: https://github.com/elastic/kibana/issues/154721

## Summary

- Fixes flaky test:
`x-pack/plugins/security_solution/cypress/e2e/detection_response/rule_management/rule_actions/bulk_actions/bulk_edit_rules_actions.cy.ts`
- Test title: `Detection rules, bulk edit of rule actions`


## Details

For: `Detection rules, bulk edit of rule actions - Restricted action
privileges - User with no privileges can't add rule actions`
- Since this test logs in with a user with missing privileges, the
"Missing privileges" callout is shown at the top of the Rules Table. The
`selectNumberOfRules();` command selects rules one by one by clicking on
their checkboxes. However, flakiness was caused when the callout was
rendered while the selection of the rules was happening, causing a
layout shift that caused the selection of a checkbox to lose focus, and
not being able to be checked. This was solved by waiting the callout to
be rendered before the selection of rules start, with the new
`waitForCallOutToBeShown`method.

For: `Detection rules, bulk edit of rule actions - All actions
privileges - before/beforeEach Clause`
- Tests were failing in the `beforeEach` clause because the first test,
mentioned above, would be logged in with a `ROLES.hunter_no_actions`
role, and logging in with a user with permissions happened in a `before`
clause instead of a `beforeEach` clause. This caused the rest of the
suite to continue with a role without permissions, and the setup of the
second test would fail as the API requests done would fail with `401`.
Moving the initial logging-in from a `before` clause to a `beforeEach`
clause solved this issue.

For: `Detection rules, bulk edit of rule actions - All actions
privileges - Add a rule action to rules (existing connector)`
- This flakiness was extremely rare, but could be reproduced after about
400 iterations. It was caused by a similar reason as the first case
above: while rules were being selected one by one, the table would auto
refresh and focus would be lost from the checkbox that was about to be
selected. This aws fixed by disabling autorefresh in the setup.

### Other changes
- Prevents the installation of `security_detection_engine` package and
creates mock rules instead.
- Creates the `waitForCallOutToBeShown` method and moves the callout IDs
spread across different files to a a single file where they are exported
from.