[EventLog] don't use legacy template APIs during initialization
resolves https://github.com/elastic/kibana/issues/164731

This removes the check for a legacy index template for the event log
for the current version it wants to create.  This was presumably done
this way, to allow for the switch from legacy to component templates.

But it's not really needed, since each stack version creates new index
templates.  So, removing the check should be fine.