Revert integration test config changes
I didn't quite understand what these SSL changes were doing, and
although they seem to be unused I'm just not going to touch them for
now. This was committed in the course of running some things locally
against a non-SSL kibana, but that time has passed.

Additionally, this reverts the disabling of observability in integration
tests. While this was useful for developing locally, it's not something
that should be in main.