[data views] fix fields for wildcard integration test (#164772)
I'm waiting for this to fail CI - this means that a new ES snapshot is
being used. Then we update the test snapshot and should be good to
merge.

The lastest ES snapshot is returning field types in a different order
which really doesn't matter to this endpoint so I'll just update the
snapshot.

Closes https://github.com/elastic/kibana/issues/164753