[Enterprise Search] fix: try in console links for code box (#164766)
## Summary

Fixed the Code box component to take the console request instead of
re-using the code snippet which is specific for the language and won't
work in our dev console.

Since the console snippets are the same I moved it to the shared project
so it can be used by both dedicated and serverless.

(cherry picked from commit 5136714f6fcd0992c275cd63b9b72bcb9d285008)