Construct HTTP log message only if needed
Previously we have unconditionally created a debug log message that
potentially involved serializing a large response object only to ignore
it if the the log level 'debug' is disabled for that logger. With this
commit we add a check if the respective level is enabled on the logger
before we proceed to construct the log message.