[Security Solution] Discover Timeline Integration URL Sync + Fixes (#163305)
This PR looks at adding below functionalities to Discover Timeline
Integration:

## Discover state URL Sync.
It changes the url sync strategy ( based on whether the discover is
running in `standalone` or `embedded` mode ) of discover so that is easy
for security solution to read/change the global/local state in the url.

|Before|After|
|--|--|

|`http://localhost:5601/app/discover?#/?_g=(filters:!(),refreshInterval:(pause:!t,value:60000),time:(from:now-15m,to:now))&_a=(columns:!(),filters:!(),index:security-solution-default,interval:auto,query:(language:kuery,query:''),sort:!(!('@timestamp',desc)))`|`http://localhost:5601/app/security/alerts?_g=(filters:!(),refreshInterval:(pause:!t,value:60000),time:(from:now-15m,to:now))&_a=(columns:!(),filters:!(),index:security-solution-default,interval:auto,query:(language:kuery,query:%27%27),sort:!(!(%27@timestamp%27,desc)))`|

@davismcphee
As an alternative, I wanted to use `customization fwk` for this but we
are in catch-22 situation where we need `stateContainer` to define a
customization but we want to do customization while defining the
`stateContainer`. Let me know if you any ideas/thoughts regarding this.

## Lens visualizations UI actions customization
Lens visualization has triggers for certain ui actions. Below are those
triggers:
1. Brush
2. Single Value Filter
3. Multiple Value Filter
4.  ⚠️ `tableRowContextMenuClick` - Not sure what is this.

https://github.com/elastic/kibana/blob/48b7acfd2427d9927d4a62ca658592c4b7a5bbb2/src/plugins/visualizations/public/embeddable/events.ts#L24-L31

This triggers are bound to the global data services, which causes issue
when triggered in security solution. Lens provided a way to override
these `triggers` via component props. This PR use customization fwk to
override and pass those props. Below videos shows the demo.

https://github.com/elastic/kibana/assets/7485038/29034f5b-f0f3-4753-a291-f88ecaabb8e6

## Discover state persistant in Security Solution
- This PR adds discover's app state it in Security solution Redux.
- [x] TODO - Syncing with local storage remaining. Will commit soon in
this PR.

## Unit / Cypress Tests
- Added unit tests + cypress tests
- [x] TODO  - Some are still in progress.
- ⚠️ TODO  - Copy these cypress tests to serverless environment.

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>
(cherry picked from commit edef2e2d40a1589f8f1cf1ce45b9994811ef0130)