[8.9] Provide glibc <3.4.20 compatible version of re2 ARM64 binary
This updates the download for our arm64 re2 binary to one with support
for older glibc versions, fixing an issue where Kibana does not start on
CentOS/RHEL 7.

The binary is built from our pipeline at
https://github.com/elastic/kibana-custom-nodejs-builds/, where re2 is
now built in a CentOS 7 container.  A download redirect will be
updated in parallel with this pull request.