[Fleet] Fix wrong fleet server host URL in kubernetes manifest yaml when policy specifies non-default fleet server (#165127)
## Summary

If an agent policy used a fleet server host which is not the default
fleet server host, then when adding an agent the kubernetes manifest
would contain the incorrect URL.

This is because the settings request was returning all fleet servers not
the one for the policy, and we were picking the first one. I have opted
instead to pass the fleet server host down to the component from the
already filtered list we have higher up the component stack, we do the
same for the enrollment token.

**Test steps**

- add a custom fleet server URL (in Fleet settings)
- create an agent policy, use your custom URL (under agent policy
settings)
- go to add an agent to the policy, select kubernetes on the platform
selector
- Check the manifest yaml, under ` - name: FLEET_URL` should be the
correct custom URL

(cherry picked from commit 526632740739193b00ca55a6346f65a17c1709e2)