fix(NA): only importing @kbn/dev-utils on serve if available in dev (#165357)
This is a follow up of https://github.com/elastic/kibana/pull/165311

Instead of duplicating the key we are only importing from
`@kbn/dev-utils` if the package is available. The serve file is also
load under dist mode where the devOnly dependencies are not availble and
as such we can't reliable load from them.