Fix Non exported APIs in observability plugin (#165576)
Fixes #165571

## Summary
The number of items is reduced from 15 to ~~11~~ 13. (There were some
params that it seems they are using in devdocs, such as
[RuleDetailsLocatorParams](https://github.com/elastic/kibana/blob/main/api_docs/observability.devdocs.json#L4503)
and
[SloDetailsLocatorParams](https://github.com/elastic/kibana/blob/main/api_docs/observability.devdocs.json#L4513C10-L4513C33))

**Command**
```
node scripts/build_api_docs --plugin observability --stats exports
```

The rest of the items cannot be fixed as they are used as type in other
files, I started a
[discussion](https://elastic.slack.com/archives/C5TQ33ND8/p1693822625635619)
to see how we can remove them from the report.