[Response Ops] add ignore_malformed to alerts mappings AGAIN
Resolves https://github.com/elastic/kibana/issues/161465

This is a re-do of https://github.com/elastic/kibana/pull/163414, which
we had to revert since data streams do not support `ignore_malformed`
on the `@timestamp` field.  We now specifically add `ignore_malformed: false`
for that field, and then use `ignore_malformed: true` at the index level.

This ignores malformed content globally across all allowed mapping types.
For existing alerts as data indices, the new setting is not applied
directly to the existing concrete indices but will be applied whenever
the alias rolls over and a new concrete index is created.

- Verify that after upgrading alerts indices created in an older version
to this branch, alerts continue to be written and read as expected.