Remove renderItem support
In a separate PR, we will update the API for declaring the navigation
tree to allow consumers to declare flyouts. The props for the API change
will be handled at a higher level: if a navigation item declares a
flyout, a wrapper around our `EuiCollapsibleNavItem` will defer to
another way of rendering the secondary action and enabling the flyout.