[Dashboard] Get panel description from the embeddable method (#166825)
Fixes #166773 

## Summary

Fixes a bug where panels are not showing descriptions from library
items.

Changes the panel to use the `embeddable.getDescription` method for
retrieving the description. This is necessary since by-reference panels
can store descriptions in the saved object. `getDescription` will check
for a description in the saved object attributes, but the panel
description will still take precedence. This is very similar to how we
get the panel title.