[SecuritySolution] Rename timeline-saving-related components (#166740)
## Summary

This PR prepares further work on the timeline-saving-related components.
As a first step we're only renaming the components to make it easier to
reason about their function.

> `SaveTimelineButton` -> `EditTimelineButton`:

We might have a dedicated `save` button more prominently in the UI in
the near future. The former "save" button actually opened up a modal
with the `edit` form. In cypress tests, the component was already
referred to as the `edit` button.

> `TimelineTitleAndDescription` -> `EditTimelineModal`:

The original name did not make it clear that it's actually a form in a
modal.

### Checklist

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios