[8.10] Disable serverless tests (#166972)
Serverless tests are only needed on the main branch. This manually
disables the 8.10 tests, and we can follow up with a branch check on
main as the long term solution.

This is also a workaround for pull requests not using cypress tests,
impacted by test failures caused by an issue between Chrome 117 and
cypress < 12.15.0. The long term solution for that is backporting
#165869.