[ES|QL] Fixes error in new metric when breakdown is a number (#167018)
## Summary

In ES|QL we don't have field formatters. We just have the type of the
field (number/date etc).

If you are in Discover and write a query which returns only one
column/row then the new metric is suggested.

<img width="1990" alt="image"
src="https://github.com/elastic/kibana/assets/17003240/79650d4b-f0c9-476b-b050-617653ccc993">

If for whatever reason the user goes to breakdown and selects the same
number variable (try_new) the chart will fail to render and you will see
in the console an error.

This happens because EC wait the title to be string but here as we don't
have field formatters this is number and fails.