Reduce the number of persisted risk inputs from 10 to 5
This has a demonstrable impact on both search and persistence times, so
decreasing this value means that risk score documents are both smaller
and quicker to generate. While this comes at the cost of decreased
first-order explainability of a risk score, the subsequent analysis of
an entity will allow investigation of ALL risk inputs in either case, so
the difference between 5 and 10 is minimal.